// src/api/generated/NodeStatus.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { NodeMode } from "./NodeMode";
import type { NodeState } from "./NodeState";

export type NodeStatus = { mode: NodeMode, state: NodeState, generation: string, height: string | null, rpc_height?: string, target_height: string | null, reachable: boolean, ready: boolean, offline: boolean, untrusted: boolean, };